re-organize stdnet files to mirror std::os::unix::net
