+++ src/sys/windows/uds/listener.rs
@@ -13,8 +13,7 @@ pub(crate) fn bind(path: &Path) -> io::Result<net::UnixListener> {
 
 pub(crate) fn accept(listener: &net::UnixListener) -> io::Result<(UnixStream, SocketAddr)> {
     listener.set_nonblocking(true)?;
-    let es = listener.accept().map(|(stream, addr)| (UnixStream::from_std(stream), addr));
-    es
+    listener.accept().map(|(stream, addr)| (UnixStream::from_std(stream), addr))
 }
 
 pub(crate) fn local_addr(listener: &net::UnixListener) -> io::Result<SocketAddr> {
